Moved run and show_menu to main.rb
